requirements.txt
python-dotenv==1.0.0
langchain==0.0.354
langchain-core==0.1.18
langchain-community==0.0.19
langchain-huggingface==0.0.2
langgraph==0.0.28
langchain-chroma==0.0.6
streamlit==1.29.0
faiss-cpu==1.7.4
pypdf==3.17.1
chromadb==0.4.22
pillow==10.1.0
docker==6.1.3
pydantic==2.5.3
sentence-transformers==2.2.2